move static vector
